requirements.txt
anthropic==0.44.0
python-dotenv==1.0.1
